restructure: models into package
