+++ main.py
@@ -1,7 +1,8 @@
 import os
 from departments_app import create_app, db
 from flask_migrate import Migrate
-from departments_app.models import Department, Employee
+from models.department import Department
+from models.employee import Employee
 
 app = create_app(os.getenv('FLASK_CONFIG') or 'default')
